requirements.txt
beeai-framework[rag,mcp]
langchain-postgres
psycopg[binary]
python-dotenv
pytest
pytest-asyncio
pytest-mock